requirements.txt
Jinja2==2.6
Markdown==2.2.1
PIL==1.1.7
PyYAML==3.10
distribute==0.6.31
html5lib==0.95
ipdb==0.7
ipython==0.13.1
mdx-smartypants==1.3
namedentities==1.301
pyPdf==1.13
readline==6.2.4.1
reportlab==2.6
wsgiref==0.1.2
-e git://github.com/chrisglass/xhtml2pdf.git@541c5357e34453663e885dd3bcf58e924d23ca2c#egg=xhtml2pdf-dev
